rails4: merges commits done at master branch into rails4 branch

+++ spec/active_record/connection_adapters/oracle_enhanced_schema_statements_spec.rb
@@ -931,6 +931,7 @@ end
       schema_define do
         create_table :test_posts, :force => true do |t|
           t.string :title, :null => false
+          t.string :content
         end
       end
       class ::TestPost < ActiveRecord::Base; end
@@ -992,10 +993,30 @@ end
       TestPost.columns_hash['title'].should be_nil
     end
 
+    it "should remove column when using change_table" do
+      schema_define do
+        change_table :test_posts do |t|
+          t.remove :title
+        end
+      end
+      TestPost.reset_column_information
+      TestPost.columns_hash['title'].should be_nil
+    end
+
+    it "should remove multiple columns when using change_table" do
+      schema_define do
+        change_table :test_posts do |t|
+          t.remove :title, :content
+        end
+      end
+      TestPost.reset_column_information
+      TestPost.columns_hash['title'].should be_nil
+      TestPost.columns_hash['content'].should be_nil
+    end
   end
 
   describe 'virtual columns in create_table' do
-    before(:all) do
+    before(:each) do
       pending "Not supported in this database version" unless @oracle11g
     end
 
@@ -1045,11 +1066,8 @@ end
   end
 
   describe 'virtual columns' do
-    before(:all) do
-      pending "Not supported in this database version" unless @oracle11g
-    end
-
     before(:each) do
+      pending "Not supported in this database version" unless @oracle11g
       expr = "( numerator/NULLIF(denominator,0) )*100"
       schema_define do
         create_table :test_fractions, :force => true do |t|
@@ -1069,8 +1087,10 @@ end
     end
 
     after(:each) do
-      schema_define do
-        drop_table :test_fractions
+      if @oracle11g
+        schema_define do
+          drop_table :test_fractions
+        end
       end
     end
